requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
websockets>=12.0
openai>=1.50.0
python-dotenv>=1.0.0

# Database
asyncpg>=0.29.0

# Auth
python-jose[cryptography]>=3.3.0
httpx>=0.27.0
